+++ CHANGELOG.md
@@ -1,6 +1,35 @@
 # Changelog
 
-## Unreleased
+## [0.4.0] - 2021-08-26
+
+This release of Garble adds support for Go 1.17.x while maintaining support for
+Go 1.16.x. A few other noteworthy changes are included:
+
+* Support obfuscating literals in more edge cases with `-literals`
+* Improve detection of `reflect` usage with standard library APIs
+* Names exported for cgo are no longer obfuscated
+* Avoid breaking consts using `iota` with `-literals`
+
+Known bugs:
+
+* obfuscating the entire standard library with `GOPRIVATE=*` is not well supported yet
+
+## [0.3.0] - 2021-05-31
+
+This release of Garble fixes a number of bugs and improves existing features,
+while maintaining support for Go 1.16.x. Notably:
+
+* Make builds reproducible even when cleaning `GOCACHE`
+* Detecting types used with reflection is more reliable
+* Cross builds with `GOPRIVATE=*` are now supported
+* Support conversion between struct types from different packages
+* Improve support for type aliases
+* Function names used with `go:linkname` are now obfuscated
+* `garble reverse` can now reverse field names and lone filenames
+
+Known bugs:
+
+* obfuscating the entire standard library with `GOPRIVATE=*` is not well supported yet
 
 ## [0.2.0] - 2021-04-08
 
@@ -39,5 +68,7 @@ Known bugs:
 * obfuscating the standard library with `GOPRIVATE=*` is not well supported yet
 * `garble test` is temporarily disabled, as it is currently broken
 
+[0.4.0]: https://github.com/burrowers/garble/releases/tag/v0.4.0
+[0.3.0]: https://github.com/burrowers/garble/releases/tag/v0.3.0
 [0.2.0]: https://github.com/burrowers/garble/releases/tag/v0.2.0
 [0.1.0]: https://github.com/burrowers/garble/releases/tag/v0.1.0